requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
supabase>=2.0.0
openai
groq
anthropic
httpx>=0.25.0
websockets
cachetools
python-jose[cryptography]
tenacity
pydantic>=2.5.0
pydantic-settings
asyncio
pytest>=7.4.0
pytest-asyncio>=0.21.0
python-multipart
stripe
python-dotenv>=1.0.0
